fix(condo): DOMA-10096 first class support
