+++ apps/condo/package.json
@@ -95,7 +95,6 @@
     "csv-stringify": "^6.2.0",
     "cuid": "^2.1.8",
     "dayjs": "^1.10.6",
-    "dd-trace": "4.46.0",
     "echarts": "^5.1.2",
     "echarts-for-react": "3.0.2",
     "emotion": "^10.0.27",
