+++ package.json
@@ -71,7 +71,8 @@
     "@sentry/profiling-node": "^7.102.1",
     "bull": "^4.8.4",
     "bull-repl": "^0.27.2",
-    "commitlint-plugin-function-rules": "^1.3.2"
+    "commitlint-plugin-function-rules": "^1.3.2",
+    "dd-trace": "4.46.0"
   },
   "husky": {
     "hooks": {
